chore: udpate

+++ contracts/test/base/BaseTest.t.sol
@@ -16,6 +16,7 @@
 
 pragma solidity ^0.8.27;
 
+import {Initializable} from "@openzeppelin/contracts/proxy/utils/Initializable.sol";
 import {EntryPoint} from "@/contracts/core/EntryPoint.sol";
 import {LightPaymaster} from "@/contracts/LightPaymaster.sol";
 import {LightTimelockController} from "@/contracts/LightTimelockController.sol";
@@ -42,8 +43,8 @@ abstract contract BaseTest is Test {
     // Initialized Event from `LightWallet.sol`
     event LightWalletInitialized(address entrypoint, bytes32 imageHash);
 
-    // Initialzed Event from `Initializable.sol` https://github.com/OpenZeppelin/openzeppelin-contracts/blob/e50c24f5839db17f46991478384bfda14acfb830/contracts/proxy/utils/Initializable.sol#L73
-    event Initialized(uint8 version);
+    // Initialzed Event from `Initializable.sol` https://github.com/OpenZeppelin/openzeppelin-contracts/blob/3e6c86392c97fbc30d3d20a378a6f58beba08eba/contracts/proxy/utils/Initializable.sol#L92
+    event Initialized(uint64 version);
 
     // ImageHashUpdated Event from `IModuleAuth.sol` https://github.com/0xsequence/wallet-contracts/blob/e0c5382636a88b4db4bcf0a70623355d7cd30fb4/contracts/modules/commons/interfaces/IModuleAuth.sol#L9
     event ImageHashUpdated(bytes32 imageHash);
